+++ docusaurus/src/pages/index.js
@@ -8,20 +8,21 @@ import HomepageWhatYouGet from '@site/src/components/HomepageWhatYouGet';
 import HomepageCommunity from '@site/src/components/HomepageCommunity';
 import HomepageFAQ from '@site/src/components/HomepageFAQ';
 import HomepageFastkafkaChat from '@site/src/components/HomepageFastkafkaChat';
+import RobotFooterIcon from '@site/src/components/RobotFooterIcon';
 
 import styles from './index.module.css';
 
 function HomepageHeader() {
-  const {siteConfig} = useDocusaurusContext();
   return (
     <header className={clsx('hero hero--primary', styles.heroBanner)}>
       <div className="container">
-        <h1 className={styles.title}>Effortless Kafka integration for web services</h1>
-        <p className={styles.description}>Open-source framework for building asynchronous web services </p>
-        <p className={styles.description}>that interact with Kafka</p>
+        <img className={styles.heroRobot} src="img/robot-hero.svg" />
+        <p className={styles.description}>Open-source framework for building asynchronous web </p>
+        <p className={styles.description}>services that interact with Kafka</p>
+        <p className={styles.descriptionMobile}>Open-source framework for building asynchronous web services that interact with Kafka</p>
         <div className={styles.buttons}>
           <Link
-            className="button button--secondary button--lg"
+            className={clsx("button button--lg", styles.heroButton)}
             to="/docs">
               Get Started
           </Link>
@@ -35,8 +36,8 @@ export default function Home() {
   const {siteConfig} = useDocusaurusContext();
   return (
     <Layout
-      title="Effortless Kafka integration for web services"
-      description="Effortless Kafka integration for web services">
+      title={siteConfig.tagline}
+      description={siteConfig.customFields.description}>
       <HomepageHeader />
       <main>
         <HomepageFeatures />
@@ -44,6 +45,7 @@ export default function Home() {
         <HomepageWhatYouGet />
         <HomepageCommunity />
         <HomepageFAQ />
+        <RobotFooterIcon />
       </main>
     </Layout>
   );