+++ fastkafka/_testing/test_utils.py
@@ -6,18 +6,18 @@ __all__ = ['logger', 'nb_safe_seed', 'mock_AIOKafkaProducer_send', 'run_script_a
 # %% ../../nbs/004_Test_Utils.ipynb 1
 import asyncio
 import hashlib
+import platform
 import shlex
+import signal
 import subprocess  # nosec
 import unittest
 import unittest.mock
 from contextlib import contextmanager
-from datetime import datetime, timedelta
 from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import *
 
 import asyncer
-from aiokafka import AIOKafkaProducer
 from IPython.display import IFrame
 
 from .._application.app import FastKafka
@@ -69,16 +69,20 @@ async def run_script_and_cancel(
     kafka_app_name: str = "kafka_app",
     generate_docs: bool = False,
 ) -> Tuple[int, bytes]:
-    """Run script and cancel after predefined time
+    """
+    Runs a script and cancels it after a predefined time.
 
     Args:
-        script: a python source code to be executed in a separate subprocess
-        script_file: name of the script where script source will be saved
-        cmd: command to execute. If None, it will be set to 'python3 -m {Path(script_file).stem}'
-        cancel_after: number of seconds before sending SIGTERM signal
+        script: A python source code to be executed in a separate subprocess.
+        script_file: Name of the script where script source will be saved.
+        cmd: Command to execute. If None, it will be set to 'python3 -m {Path(script_file).stem}'.
+        cancel_after: Number of seconds before sending SIGTERM signal.
+        app_name: Name of the app.
+        kafka_app_name: Name of the Kafka app.
+        generate_docs: Flag indicating whether to generate docs.
 
     Returns:
-        A tuple containing exit code and combined stdout and stderr as a binary string
+        A tuple containing the exit code and combined stdout and stderr as a binary string.
     """
     if script_file is None:
         script_file = "script.py"
@@ -106,17 +110,38 @@ async def run_script_and_cancel(
                     f"Generating docs failed for: {Path(script_file).stem}:{kafka_app_name}, ignoring it for now."
                 )
 
-        proc = subprocess.Popen(  # nosec: [B603:subprocess_without_shell_equals_true] subprocess call - check for execution of untrusted input.
-            shlex.split(cmd), stdout=subprocess.PIPE, stderr=subprocess.STDOUT, cwd=d
+        creationflags = 0 if platform.system() != "Windows" else subprocess.CREATE_NEW_PROCESS_GROUP  # type: ignore
+        proc = subprocess.Popen(
+            shlex.split(cmd),
+            stdout=subprocess.PIPE,
+            stderr=subprocess.STDOUT,
+            cwd=d,
+            shell=True  # nosec: [B602:subprocess_without_shell_equals_true] subprocess call - check for execution of untrusted input.
+            if platform.system() == "Windows"
+            else False,
+            creationflags=creationflags,
         )
         await asyncio.sleep(cancel_after)
-        proc.terminate()
+        if platform.system() == "Windows":
+            proc.send_signal(signal.CTRL_BREAK_EVENT)  # type: ignore
+        else:
+            proc.terminate()
         output, _ = proc.communicate()
 
         return (proc.returncode, output)
 
 # %% ../../nbs/004_Test_Utils.ipynb 14
 async def display_docs(docs_path: str, port: int = 4000) -> None:
+    """
+    Serves the documentation using an HTTP server.
+
+    Args:
+        docs_path: Path to the documentation.
+        port: Port number for the HTTP server. Defaults to 4000.
+
+    Returns:
+        None
+    """
     with change_dir(docs_path):
         process = await asyncio.create_subprocess_exec(
             "python3",